refactor: support generating JSON schema & OpenAPI from dynamic TS code
