+++ common/package.json
@@ -6,6 +6,7 @@
   "scripts": {},
   "dependencies": {
     "@saltify/milky-protocol": "workspace:*",
+    "sucrase": "^3.35.1",
     "zod": "^4.3.6"
   },
   "devDependencies": {
