+++ common/src/index.ts
@@ -1,6 +1,8 @@
 export { generateDartJsonSerializableSpec } from './generator/dart/json_serializable';
+export { generateJsonSchema } from './generator/json-schema';
 export { generateKotlinxSerializationSpec } from './generator/kotlin/kotlinx-serialization';
 export { generateMarkdownRoadmap } from './generator/markdown/roadmap';
+export { generateOpenApiSpec } from './generator/openapi';
 export { generateRustSerdeSpec } from './generator/rust/serde';
 export { generateTypeScriptStaticSpec } from './generator/typescript/static';
 export { generateTypeScriptZodSpec } from './generator/typescript/zod';
