+++ docs/app/raw/json-schema/schema.json/route.ts
@@ -1,21 +1,8 @@
-import {} from '@/app/raw/registry';
-import * as t from '@saltify/milky-types';
-import z from 'zod';
-
-function generateJsonSchema() {
-  return {
-    milkyVersion: t.milkyVersion,
-    packageVersion: t.milkyPackageVersion,
-    schemas: z.toJSONSchema(z.globalRegistry, {
-      metadata: z.globalRegistry,
-      io: 'input',
-    }).schemas,
-  };
-}
-
+import { ir } from '@saltify/milky-protocol';
+import { generateJsonSchema } from '@saltify/milky-common';
 
 export const dynamic = 'force-static';
 
-export function GET() {
-  return new Response(JSON.stringify(generateJsonSchema()));
+export async function GET() {
+  return new Response(JSON.stringify(await generateJsonSchema(ir, await import('@saltify/milky-types'))));
 }
