+++ docs/app/raw/rust/serde.txt/route.ts
@@ -1,5 +1,5 @@
-import { generateRustSerdeSpec } from '@saltify/milky-common/src/generator/rust/serde';
 import { ir } from '@saltify/milky-protocol';
+import { generateRustSerdeSpec } from '@saltify/milky-common';
 
 export const dynamic = 'force-static';
